refactor(qa): split Mantis evidence artifact handling
