+++ extensions/qa-lab/src/mantis/run.runtime.ts
@@ -3,8 +3,13 @@ import fs from "node:fs/promises";
 import path from "node:path";
 import { formatErrorMessage } from "openclaw/plugin-sdk/error-runtime";
 import { ensureRepoBoundDirectory, resolveRepoRelativeOutputDir } from "../cli-paths.js";
-import { QA_EVIDENCE_FILENAME, validateQaEvidenceSummaryJson } from "../evidence-summary.js";
 import { trimToValue } from "../mantis-options.runtime.js";
+import {
+  copyMantisDirContents,
+  copyMantisLaneArtifact,
+  readMantisLaneResult,
+  type LaneResult,
+} from "./run-artifacts.runtime.js";
 import {
   assertMantisCommandNotAborted,
   createMantisWorktreeDirectory,
@@ -45,33 +50,6 @@ type MantisBeforeAfterResult = {
   outputDir: string;
   reportPath: string;
   status: "pass" | "fail";
-};
-
-type DiscordQaSummary = {
-  scenarios?: {
-    artifactPaths?: Record<string, string>;
-    details?: string;
-    id?: string;
-    status?: string;
-    title?: string;
-  }[];
-};
-
-type NormalizedScenarioSummary = {
-  details?: string;
-  screenshotPath?: string;
-  status: string;
-  summaryPath: string;
-  videoPath?: string;
-};
-
-type LaneResult = {
-  outputDir: string;
-  scenarioDetails?: string;
-  screenshotPath?: string;
-  status: string;
-  summaryPath: string;
-  videoPath?: string;
 };
 
 type MantisScenarioConfig = {
@@ -158,75 +136,6 @@ function normalizeRequiredLiteral<T extends string>(
 function defaultOutputDir(repoRoot: string, startedAt: Date) {
   const stamp = startedAt.toISOString().replace(/[:.]/gu, "-");
   return path.join(repoRoot, ".artifacts", "qa-e2e", "mantis", `run-${stamp}`);
-}
-
-async function copyDirContents(sourceDir: string, targetDir: string) {
-  await fs.rm(targetDir, { force: true, recursive: true });
-  await fs.mkdir(targetDir, { recursive: true });
-  await fs.cp(sourceDir, targetDir, { recursive: true });
-}
-
-async function readLaneResult(params: {
-  laneOutputDir: string;
-  publishedLaneDir: string;
-  scenario: string;
-}) {
-  const normalized = await readNormalizedLaneResult(params);
-  if (normalized) {
-    return {
-      outputDir: params.publishedLaneDir,
-      scenarioDetails: normalized.details,
-      screenshotPath: normalized.screenshotPath,
-      status: normalized.status,
-      summaryPath: normalized.summaryPath,
-      videoPath: normalized.videoPath,
-    } satisfies LaneResult;
-  }
-
-  const summaryPath = path.join(params.publishedLaneDir, "discord-qa-summary.json");
-  const summary = JSON.parse(await fs.readFile(summaryPath, "utf8")) as DiscordQaSummary;
-  const scenarioSummary =
-    summary.scenarios?.find((entry) => entry.id === params.scenario) ?? summary.scenarios?.[0];
-  const status = scenarioSummary?.status ?? "fail";
-  const screenshotPath = scenarioSummary?.artifactPaths?.screenshot;
-  const videoPath = scenarioSummary?.artifactPaths?.video;
-  return {
-    outputDir: params.publishedLaneDir,
-    scenarioDetails: scenarioSummary?.details,
-    screenshotPath,
-    status,
-    summaryPath,
-    videoPath,
-  } satisfies LaneResult;
-}
-
-async function readNormalizedLaneResult(params: {
-  publishedLaneDir: string;
-  scenario: string;
-}): Promise<NormalizedScenarioSummary | undefined> {
-  const summaryPath = path.join(params.publishedLaneDir, QA_EVIDENCE_FILENAME);
-  let rawSummary: string;
-  try {
-    rawSummary = await fs.readFile(summaryPath, "utf8");
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code === "ENOENT") {
-      return undefined;
-    }
-    throw error;
-  }
-
-  const summary = validateQaEvidenceSummaryJson(JSON.parse(rawSummary));
-  const entry =
-    summary.entries.find((candidate) => candidate.test.id === params.scenario) ??
-    summary.entries[0];
-  const artifacts = entry?.execution?.artifacts ?? [];
-  return {
-    details: entry?.result.failure?.reason,
-    screenshotPath: artifacts.find((artifact) => artifact.kind === "screenshot")?.path,
-    status: entry?.result.status ?? "fail",
-    summaryPath,
-    videoPath: artifacts.find((artifact) => artifact.kind === "video")?.path,
-  };
 }
 
 function renderReport(params: {
@@ -390,30 +299,6 @@ function buildEvidenceManifest(params: {
   };
 }
 
-async function copyScreenshot(params: { lane: "baseline" | "candidate"; result: LaneResult }) {
-  if (!params.result.screenshotPath) {
-    return undefined;
-  }
-  const source = path.isAbsolute(params.result.screenshotPath)
-    ? params.result.screenshotPath
-    : path.join(params.result.outputDir, params.result.screenshotPath);
-  const target = path.join(params.result.outputDir, `${params.lane}.png`);
-  await fs.copyFile(source, target);
-  return target;
-}
-
-async function copyVideo(params: { lane: "baseline" | "candidate"; result: LaneResult }) {
-  if (!params.result.videoPath) {
-    return undefined;
-  }
-  const source = path.isAbsolute(params.result.videoPath)
-    ? params.result.videoPath
-    : path.join(params.result.outputDir, params.result.videoPath);
-  const target = path.join(params.result.outputDir, `${params.lane}.mp4`);
-  await fs.copyFile(source, target);
-  return target;
-}
-
 async function runLane(params: {
   lane: "baseline" | "candidate";
   outputDir: string;
@@ -549,14 +434,22 @@ async function runLane(params: {
       runner: params.runner,
     });
     const publishedLaneDir = path.join(params.outputDir, params.lane);
-    await copyDirContents(path.join(worktreeDir, worktreeOutputDir), publishedLaneDir);
-    const result = await readLaneResult({
+    await copyMantisDirContents(path.join(worktreeDir, worktreeOutputDir), publishedLaneDir);
+    const result = await readMantisLaneResult({
       laneOutputDir: path.join(worktreeDir, worktreeOutputDir),
       publishedLaneDir,
       scenario: params.scenario,
     });
-    const copiedScreenshot = await copyScreenshot({ lane: params.lane, result });
-    const copiedVideo = await copyVideo({ lane: params.lane, result });
+    const copiedScreenshot = await copyMantisLaneArtifact({
+      kind: "screenshot",
+      lane: params.lane,
+      result,
+    });
+    const copiedVideo = await copyMantisLaneArtifact({
+      kind: "video",
+      lane: params.lane,
+      result,
+    });
     laneResult = {
       ...result,
       screenshotPath: copiedScreenshot ?? result.screenshotPath,
